Add uid identified to commandItem and integrate it into CommandNode, this fixes deletion bug and allows adding multiple elements with same names

+++ src/main/java/com/ttzv/item/ui/CommandBoxScn.java
@@ -2,6 +2,7 @@ package com.ttzv.item.ui;
 
 import com.ttzv.item.entity.CommandItem;
 import com.ttzv.item.entity.EntityDAO;
+import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.collections.ObservableListBase;
@@ -10,6 +11,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
+import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
@@ -19,9 +21,8 @@ import javax.naming.NamingException;
 import java.io.IOException;
 import java.security.GeneralSecurityException;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
+import java.util.stream.Collectors;
 
 public class CommandBoxScn extends AnchorPane {
 
@@ -29,16 +30,25 @@ public class CommandBoxScn extends AnchorPane {
     private TextField searchField;
 
     @FXML
-    private VBox commandList;
+    private ScrollPane scrollPane;
 
+    @FXML
+    private VBox commandList;
 
     @FXML
     void addAction(ActionEvent event) {
+        boolean lastAddedNodeTitleIsEmpty = commandNodelist.get(commandNodelist.size() - 1).getTitle().isEmpty();
+        if (!lastAddedNodeTitleIsEmpty) {
             CommandNode commandNode = new CommandNode();
+            commandNode.setUid(getNextUid());
             commandNodelist.add(commandNode);
             addDeleteBtnAction(commandNode);
             addUpdateBtnAction(commandNode);
-        refreshList();
+            refreshList();
+            scrollToBottom();
+        } else {
+            scrollToBottom();
+        }
     }
 
 
@@ -51,15 +61,31 @@ public class CommandBoxScn extends AnchorPane {
         addSearchFieldListener();
     }
 
+    @FXML
+    void refresh(ActionEvent event) {
+        try {
+            commandItemList = entityDAOcmddb.getAllEntities();
+        } catch (SQLException | IOException | NamingException | GeneralSecurityException e) {
+            e.printStackTrace();
+        }
+        commandNodelist.clear();
+        for (CommandItem c : commandItemList) {
+            addCommandNode(c);
+        }
+        refreshList();
+    }
 
     private EntityDAO<CommandItem> entityDAOcmddb;
     private List<CommandItem> commandItemList;
+    private List<Integer> commandItemUidList;
     private ObservableList<CommandNode> commandNodelist;
+
 
 
     public CommandBoxScn(EntityDAO<CommandItem> entityDAOcmddb) throws SQLException, NamingException, GeneralSecurityException, IOException {
         this.entityDAOcmddb = entityDAOcmddb;
         commandItemList = entityDAOcmddb.getAllEntities();
+        commandItemUidList = commandItemList.stream().map(c -> Integer.parseInt(c.getUid())).collect(Collectors.toList());
         commandNodelist = FXCollections.observableList(new ArrayList<>());
 
         FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("/fxml/commandboxscn.fxml"));
@@ -76,7 +102,7 @@ public class CommandBoxScn extends AnchorPane {
         commandList.getChildren().setAll(commandNodelist);
     }
 
-    private void deleteitem(String title){
+    private void deleteitem(String uid){
         Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
         alert.setTitle("Potwierdzenie usunięcia");
         alert.setHeaderText("Czy na pewno chcesz usunąć ten element?");
@@ -84,40 +110,42 @@ public class CommandBoxScn extends AnchorPane {
         Optional<ButtonType> result = alert.showAndWait();
 
         if(result.get() == ButtonType.OK) {
-            CommandItem commandItem = new CommandItem(title);
-            try {
-                entityDAOcmddb.deleteEntity(commandItem);
-            } catch (SQLException | IOException ex) {
-                ex.printStackTrace();
+            CommandItem commandItem = new CommandItem(uid);
+            if(commandItemUidList.contains(Integer.parseInt(commandItem.getUid()))) {
+                try {
+                    entityDAOcmddb.deleteEntity(commandItem);
+                } catch (SQLException | IOException ex) {
+                    ex.printStackTrace();
+                }
             }
-            commandNodelist.removeIf(p -> p.getTitle().equals(title));
+            commandNodelist.removeIf(p -> p.getUid().equals(uid));
             refreshList();
         }
     }
 
     private void addSearchFieldListener(){
         searchField.textProperty().addListener((observableValue, s, t1) -> {
-            commandList.getChildren().setAll(commandNodelist.filtered(c -> c.getTitle().contains(t1)));
+            commandList.getChildren().setAll(commandNodelist.filtered(c -> c.getTitle().contains(t1.toLowerCase().trim()) || c.getTagsFieldText().contains(t1.toLowerCase().trim())));
         });
     }
 
     private void addDeleteBtnAction(CommandNode commandNode){
         commandNode.getBtnDelete().setOnAction(e -> {
-            String title = commandNode.getTitle();
-            if(!title.isEmpty()) {
-                deleteitem(title);
-            }
+            String uid = commandNode.getUid();
+            deleteitem(uid);
         });
     }
 
     private void addUpdateBtnAction(CommandNode commandNode){
         commandNode.getBtnUpdate().setOnAction(e -> {
-            CommandItem commandItem = new CommandItem(commandNode.getTitle(), commandNode.getContent(), commandNode.getTagsFieldText());
-            try {
-                this.entityDAOcmddb.updateEntity(commandItem);
-            } catch (SQLException | IOException ex) {
-                ex.printStackTrace();
-            }
+            CommandItem commandItem = new CommandItem(commandNode.getUid(), commandNode.getTitle().trim(), commandNode.getContent().trim(), commandNode.getTagsFieldText().trim());
+                try {
+                    this.entityDAOcmddb.updateEntity(commandItem);
+                } catch (SQLException | IOException ex) {
+                    ex.printStackTrace();
+                }
+                commandItemList.add(commandItem);
+                commandItemUidList.add(Integer.parseInt(commandItem.getUid()));
         });
     }
 
@@ -126,17 +154,21 @@ public class CommandBoxScn extends AnchorPane {
         commandNode.setTitleField(commandItem.getCommandTitle());
         commandNode.setContent(commandItem.getCommandContents());
         commandNode.setTagsField(commandItem.getTags());
+        commandNode.setUid(commandItem.getUid());
         addDeleteBtnAction(commandNode);
         addUpdateBtnAction(commandNode);
         this.commandNodelist.add(commandNode);
+        this.commandItemUidList.add(Integer.parseInt(commandItem.getUid()));
     }
 
-    /*
-    private CommandItem getCommandItem(String title){
+    private void scrollToBottom(){
+        Platform.runLater(() -> this.scrollPane.setVvalue(1.0));
 
     }
-    */
 
-
+    private String getNextUid(){
+        Collections.sort(commandItemUidList);
+        return String.valueOf(commandItemUidList.get(commandItemUidList.size()-1) + 1);
+    }
 
 }
